+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - [#712] Fix an issue where preview might break if certain internal objects are destroyed unexpectedly.
+- [#718] Fix NDMF preview in cross platform projects not handling cross platform avatar roots appropriately.
 
 ### Changed